move device handling to container iface
